# Change uid intersection policy in some cases