--- a/algo/uidlist.go
+++ b/algo/uidlist.go
@@ -37,6 +37,24 @@ func ApplyFilter(u *pb.List, f func(uint64, int) bool) {
 	u.Uids = out
 }
 
+func IntersectCompressedWith2(pack *pb.UidPack, afterUID uint64, v, o *pb.List) {
+	if pack == nil {
+		return
+	}
+	dec := codec.Decoder{Pack: pack}
+	dec.Seek(afterUID, codec.SeekStart)
+
+	k := &pb.List{Uids: make([]uint64, 0)}
+
+	for ; dec.Valid(); dec.Next() {
+		for _, uid := range dec.Uids() {
+			k.Uids = append(k.Uids, uid)
+		}
+	}
+
+	IntersectWith(k, v, o)
+}
+
 // IntersectCompressedWith intersects a packed list of UIDs with another list
 // and writes the output to o.
 func IntersectCompressedWith(pack *pb.UidPack, afterUID uint64, v, o *pb.List) {