--- a/dgraph/cmd/alpha/run.go
+++ b/dgraph/cmd/alpha/run.go
@@ -671,8 +671,6 @@ func run() {
 		os.Exit(1)
 	}
 
-	worker.SetConfiguration(&opts)
-
 	ips, err := getIPsFromString(security.GetString("whitelist"))
 	x.Check(err)
 
@@ -683,7 +681,7 @@ func run() {
 
 	raft := z.NewSuperFlag(Alpha.Conf.GetString("raft")).MergeAndCheckDefault(worker.RaftDefaults)
 	x.WorkerConfig = x.WorkerOptions{
-		TmpDir:              Alpha.Conf.GetString("tmp"),
+		TmpDir:              "./t",
 		ExportPath:          Alpha.Conf.GetString("export"),
 		ZeroAddr:            strings.Split(Alpha.Conf.GetString("zero"), ","),
 		Raft:                raft,
@@ -700,7 +698,9 @@ func run() {
 		Badger:              bopts,
 	}
 	x.WorkerConfig.Parse(Alpha.Conf)
+	fmt.Println("WORKER CONFIG ============", x.WorkerConfig)
 
+	worker.SetConfiguration(&opts)
 	if telemetry.GetBool("reports") {
 		go edgraph.PeriodicallyPostTelemetry()
 	}