--- a/posting/list.go
+++ b/posting/list.go
@@ -20,6 +20,7 @@ import (
 	"bytes"
 	"context"
 	"encoding/hex"
+	"fmt"
 	"log"
 	"math"
 	"sort"
@@ -1076,7 +1077,7 @@ func (l *List) encode(out *rollupOutput, readTs uint64, split bool) error {
 		return errors.Wrapf(err, "cannot iterate through the list")
 	}
 	plist.Pack = enc.Done()
-	l.CreateUnpacked()
+	//l.CreateUnpacked()
 	if plist.Pack != nil {
 		if plist.Pack.BlockSize != uint32(blockSize) {
 			return errors.Errorf("actual block size %d is different from expected value %d",
@@ -1173,8 +1174,15 @@ func (l *List) Uids(ctx context.Context, opt ListOptions) (*pb.List, error) {
 			return out, ErrTsTooOld
 		}
 		stop := x.SpanTimer(span, "IntersectCompressedWith")
-		//algo.IntersectCompressedWith(l.plist.Pack, opt.AfterUid, opt.Intersect, out)
-		algo.IntersectWithAfter(l.immutable, opt.Intersect, out, opt.AfterUid)
+		if (len(opt.Intersect.Uids) / codec.ApproxLen(l.plist.Pack)) < 1000 {
+			l.CreateUnpacked()
+			algo.IntersectWithAfter(l.immutable, opt.Intersect, out, opt.AfterUid)
+			fmt.Println("Ratio:", len(l.immutable.Uids)/len(opt.Intersect.Uids), len(opt.Intersect.Uids)/len(l.immutable.Uids))
+		} else {
+			fmt.Println("Compressed Intersect")
+			algo.IntersectCompressedWith(l.plist.Pack, opt.AfterUid, opt.Intersect, out)
+
+		}
 		stop()
 		l.RUnlock()
 		return out, nil