--- a/posting/lists.go
+++ b/posting/lists.go
@@ -168,6 +168,7 @@ func (lc *LocalCache) getInternal(key []byte, readFromDisk bool) (*List, error) 
 
 	skey := string(key)
 	if pl := lc.getNoStore(skey); pl != nil {
+		pl.CreateUnpacked()
 		return pl, nil
 	}