--- a/posting/mvcc.go
+++ b/posting/mvcc.go
@@ -462,6 +462,28 @@ func ReadPostingList(key []byte, it *badger.Iterator) (*List, error) {
 	return l, nil
 }
 
+func GetKey(key []byte, readTs uint64) (*pb.PostingList, error) {
+	txn := pstore.NewTransactionAt(readTs, false)
+	item, err := txn.Get(key)
+	if err != nil {
+		return nil, err
+	}
+	pl := &pb.PostingList{}
+
+	err = item.Value(func(val []byte) error {
+		if err := pl.Unmarshal(val); err != nil {
+			return err
+		}
+		return nil
+	})
+
+	if err != nil {
+		return nil, err
+	}
+
+	return pl, nil
+}
+
 func getNew(key []byte, pstore *badger.DB, readTs uint64) (*List, error) {
 	cachedVal, ok := lCache.Get(key)
 	if ok {