--- a/posting/mvcc.go
+++ b/posting/mvcc.go
@@ -72,6 +72,9 @@ var (
 	IncrRollup = &incrRollupi{
 		priorityKeys: make([]*pooledKeys, 2),
 	}
+
+	ReadTime      = time.Duration(0)
+	UnmarshalTime = time.Duration(0)
 )
 
 func init() {
@@ -462,15 +465,17 @@ func ReadPostingList(key []byte, it *badger.Iterator) (*List, error) {
 	return l, nil
 }
 
-func GetKey(key []byte, readTs uint64) (*pb.PostingList, error) {
+func GetKey(key []byte, readTs uint64) (*pb.PostingList, error, int) {
 	txn := pstore.NewTransactionAt(readTs, false)
 	item, err := txn.Get(key)
 	if err != nil {
-		return nil, err
+		return nil, err, 0
 	}
 	pl := &pb.PostingList{}
+	k := 0
 
 	err = item.Value(func(val []byte) error {
+		k = len(key) + len(val)
 		if err := pl.Unmarshal(val); err != nil {
 			return err
 		}
@@ -478,10 +483,10 @@ func GetKey(key []byte, readTs uint64) (*pb.PostingList, error) {
 	})
 
 	if err != nil {
-		return nil, err
+		return nil, err, 0
 	}
 
-	return pl, nil
+	return pl, nil, k
 }
 
 func getNew(key []byte, pstore *badger.DB, readTs uint64) (*List, error) {