--- a/worker/server_state.go
+++ b/worker/server_state.go
@@ -18,6 +18,7 @@ package worker
 
 import (
 	"context"
+	"fmt"
 	"math"
 	"os"
 	"time"
@@ -118,7 +119,9 @@ func (s *ServerState) initStorage() {
 
 	{
 		// Write Ahead Log directory
-		x.Checkf(os.MkdirAll(Config.WALDir, 0700), "Error while creating WAL dir.")
+		if Config.WALDir != "" {
+			x.Checkf(os.MkdirAll(Config.WALDir, 0700), "Error while creating WAL dir.")
+		}
 		s.WALstore, err = raftwal.InitEncrypted(Config.WALDir, x.WorkerConfig.EncryptionKey)
 		x.Check(err)
 	}
@@ -126,11 +129,19 @@ func (s *ServerState) initStorage() {
 		// Postings directory
 		// All the writes to posting store should be synchronous. We use batched writers
 		// for posting lists, so the cost of sync writes is amortized.
-		x.Check(os.MkdirAll(Config.PostingDir, 0700))
+		fmt.Println("Here", Config.PostingDir)
 		opt := x.WorkerConfig.Badger.
-			WithDir(Config.PostingDir).WithValueDir(Config.PostingDir).
 			WithNumVersionsToKeep(math.MaxInt32).
 			WithNamespaceOffset(x.NamespaceOffset)
+
+		if Config.PostingDir != "" {
+			fmt.Println("SETTING POSTING DIR", Config.PostingDir, Config.PostingDir == "", []byte(Config.PostingDir), "<- Config")
+			x.Check(os.MkdirAll(Config.PostingDir, 0700))
+			opt = opt.WithDir(Config.PostingDir).WithValueDir(Config.PostingDir)
+		} else {
+			fmt.Println("HERE@@@@@", opt.Dir, opt.ValueDir)
+		}
+
 		opt = setBadgerOptions(opt)
 
 		// Print the options w/o exposing key.