Title format:

 - `Topic(Area): Feature`
 - `Topic` must be one of `build|ci|docs|feat|fix|perf|refactor|chore|test`
 - `Area` must be one of `acl|audit|backup|badger|cdc|dql|export|graphql|indexing|multi-tenancy|raft|restore|upgrade|zero`
 - * Add [BREAKING] if it is a breaking change

------------------
Body Format:

Description: <add description>
Fixes: <GitHub Issue>
Closes: <JIRA Issue>
Docs: <docs PR>

LDBC Results. Upto 13.46% better

```
goos: linux
goarch: amd64
pkg: ldbc/standalone/dgraph
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
               │     in1      │                  in2                   │
               │    sec/op    │    sec/op      vs base                 │
Queries/IS01-8   659.9µ ± ∞ ¹    944.9µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Queries/IS02-8   12.17m ± ∞ ¹    13.52m ± ∞ ¹        ~ (p=1.000 n=1) ²
Queries/IS03-8   833.3µ ± ∞ ¹   1219.4µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Queries/IS04-8   575.7µ ± ∞ ¹   1009.3µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Queries/IS05-8   770.8µ ± ∞ ¹   1079.8µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Queries/IS06-8   985.6µ ± ∞ ¹   1193.3µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Queries/IS07-8   4.792m ± ∞ ¹    4.810m ± ∞ ¹        ~ (p=1.000 n=1) ²
Queries/IC01-8    1.337 ± ∞ ¹     1.298 ± ∞ ¹        ~ (p=1.000 n=1) ²
Queries/IC02-8    1.308 ± ∞ ¹     1.392 ± ∞ ¹        ~ (p=1.000 n=1) ²
Queries/IC04-8    1.311 ± ∞ ¹     1.284 ± ∞ ¹        ~ (p=1.000 n=1) ²
Queries/IC06-8    2.853 ± ∞ ¹     3.035 ± ∞ ¹        ~ (p=1.000 n=1) ²
Queries/IC07-8   534.4m ± ∞ ¹    420.7m ± ∞ ¹        ~ (p=1.000 n=1) ²
Queries/IC08-8   33.02m ± ∞ ¹    32.92m ± ∞ ¹        ~ (p=1.000 n=1) ²
Queries/IC09-8    28.89 ± ∞ ¹     37.55 ± ∞ ¹        ~ (p=1.000 n=1) ²
Queries/IC10-8    35.33 ± ∞ ¹     36.77 ± ∞ ¹        ~ (p=1.000 n=1) ²
Queries/IC11-8   109.9m ± ∞ ¹    107.2m ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean          67.41m          76.60m        +13.64%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

```


Benchmark results:
```
BenchmarkListIntersectRatio/:IntersectWith:ratio=0.010000:size=100:overlap=0.01:-8                930638              1336 ns/op                                                                                                                                                                                                                                                                
BenchmarkListIntersectRatio/compressed:IntersectWith:ratio=0.010000:size=100:overlap=0.01:-8             1368980               797.5 ns/op                                                                                                                                                                                                                                                      
                                                                                                                                                                                                                                                                                                                                                                                                
len: 10, compressed: 159, bytes/int: 15.900000                                                                                                                                                                                                                                                                                                                                                  
BenchmarkListIntersectRatio/:IntersectWith:ratio=0.100000:size=100:overlap=0.01:-8                        863913              1408 ns/op                                                                                                                                                                                                                                                        
BenchmarkListIntersectRatio/compressed:IntersectWith:ratio=0.100000:size=100:overlap=0.01:-8             1411018               844.4 ns/op                                                                                                                                                                                                                                                      
                                                                                                                                                                                                                                                                                                                                                                                                
len: 100, compressed: 156, bytes/int: 1.560000                                                                                                                                                                                                                                                                                                                                                  
BenchmarkListIntersectRatio/:IntersectWith:ratio=1.000000:size=100:overlap=0.01:-8                        679086              1526 ns/op                                                                                                                                                                                                                                                        
BenchmarkListIntersectRatio/compressed:IntersectWith:ratio=1.000000:size=100:overlap=0.01:-8             1213830               982.4 ns/op                                                                                                                                                                                                                                                      
                                                                                                                                                                                                                                                                                                                                                                                                
len: 1000, compressed: 158, bytes/int: 0.158000                                                                                                                                                                                                                                                                                                                                                 
BenchmarkListIntersectRatio/:IntersectWith:ratio=10.000000:size=100:overlap=0.01:-8                       610076              1950 ns/op                                                                                                                                                                                                                                                        
BenchmarkListIntersectRatio/compressed:IntersectWith:ratio=10.000000:size=100:overlap=0.01:-8             677944              1728 ns/op                                                                                                                                                                                                                                                        
                                                                                                                                                                                                                                                                                                                                                                                                
len: 10000, compressed: 158, bytes/int: 0.015800                                                                                                                                                                                                                                                                                                                                                
BenchmarkListIntersectRatio/:IntersectWith:ratio=100.000000:size=100:overlap=0.01:-8                      404162              2945 ns/op                                                                                                                                                                                                                                                        
BenchmarkListIntersectRatio/compressed:IntersectWith:ratio=100.000000:size=100:overlap=0.01:-8             52226             23293 ns/op                                                                                                                                                                                                                                                        
                                                                                                                                                                                                                                                                                                                                                                                                
len: 20000, compressed: 157, bytes/int: 0.007850                                                                                                                                                                                                                                                                                                                                                
BenchmarkListIntersectRatio/:IntersectWith:ratio=200.000000:size=100:overlap=0.01:-8                      360571              3185 ns/op                                                                                                                                                                                                                                                        
BenchmarkListIntersectRatio/compressed:IntersectWith:ratio=200.000000:size=100:overlap=0.01:-8              4987            225165 ns/op                                                                                                                                                                                                                                                        
                                                                                                                                                                                                                                                                                                                                                                                                
len: 40000, compressed: 158, bytes/int: 0.003950                                                                                                                                                                                                                                                                                                                                                
BenchmarkListIntersectRatio/:IntersectWith:ratio=400.000000:size=100:overlap=0.01:-8                      302030              4073 ns/op                                                                                                                                                                                                                                                        
BenchmarkListIntersectRatio/compressed:IntersectWith:ratio=400.000000:size=100:overlap=0.01:-8             14863             79218 ns/op                                                                                                                                                                                                                                                        
                                                                                                                                                                                                                                                                                                                                                                                                
len: 50000, compressed: 159, bytes/int: 0.003180                                                                                                                                                                                                                                                                                                                                                
BenchmarkListIntersectRatio/:IntersectWith:ratio=500.000000:size=100:overlap=0.01:-8                      243756              4642 ns/op                                                                                                                                                                                                                                                        
BenchmarkListIntersectRatio/compressed:IntersectWith:ratio=500.000000:size=100:overlap=0.01:-8             24970             48431 ns/op                                                                                                                                                                                                                                                        
                                                                                                                                                                                                                                                                                                                                                                                                
len: 100000, compressed: 156, bytes/int: 0.001560                                                                                                                                                                                                                                                                                                                                               
BenchmarkListIntersectRatio/:IntersectWith:ratio=1000.000000:size=100:overlap=0.01:-8                     250286              4713 ns/op                                                                                                                                                                                                                                                        
BenchmarkListIntersectRatio/compressed:IntersectWith:ratio=1000.000000:size=100:overlap=0.01:-8             2589            450080 ns/op                                                                                                                                                                                                                                                        
                                                                                                                                                                                                                                                                                                                                                                                                
len: 10, compressed: 1389, bytes/int: 138.900000                                                                                                                                                                                                                                                                                                                                                
BenchmarkListIntersectRatio/:IntersectWith:ratio=0.010000:size=1000:overlap=0.01:-8                       148689              8170 ns/op                                                                                                                                                                                                                                                        
BenchmarkListIntersectRatio/compressed:IntersectWith:ratio=0.010000:size=1000:overlap=0.01:-8             296518              4018 ns/op                                                                                                                                                                                                                                                        
                                                                                                                                                                                                                                                                                                                                                                                                
len: 100, compressed: 1397, bytes/int: 13.970000                                                                                                                                                                                                                                                                                                                                                
BenchmarkListIntersectRatio/:IntersectWith:ratio=0.100000:size=1000:overlap=0.01:-8                       129889             12517 ns/op                                                                                                                                                                                                                                                        
BenchmarkListIntersectRatio/compressed:IntersectWith:ratio=0.100000:size=1000:overlap=0.01:-8             186576              6241 ns/op                                                                                                                                                                                                                                                        
                                                                                                                                                                                                                                                                                                                                                                                                
len: 1000, compressed: 1388, bytes/int: 1.388000                                                                                                                                                                                                                                                                                                                                                
BenchmarkListIntersectRatio/:IntersectWith:ratio=1.000000:size=1000:overlap=0.01:-8                        76354             15379 ns/op                                                                                                                                                                                                                                                        
BenchmarkListIntersectRatio/compressed:IntersectWith:ratio=1.000000:size=1000:overlap=0.01:-8              61672             19026 ns/op                                                                                                                                                                                                                                                        
…